requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.76

# Deep Learning for Visual Analysis
tensorflow==2.13.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.9.0

# Web Scraping (for product webpages)
beautifulsoup4==4.12.2
requests==2.31.0

# Natural Language Processing
nltk==3.8.1
spacy==3.6.1
transformers==4.33.0
sentence-transformers==2.2.2

# Scientific Computing
scipy==1.11.2

# Data Visualization (for your analysis)
matplotlib==3.7.2
seaborn==0.12.2

# Dashboard
streamlit
plotly

# PDF Report Generation
reportlab

# Progress Bars (to see progress while running)
tqdm==4.66.1

# API Access (for PubMed, EU Register)
biopython==1.81

# Anomaly Detection
pyod==1.1.0
